server/products/pyproject.toml
[tool.poetry]
name = "products"
version = "0.1.0"
description = "products service"
authors = ["bananichdev <dimitri-05@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

fastapi = "^0.110.1"
uvicorn = "^0.29.0"

sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"

httpx = "^0.27.0"

python-dotenv = "^1.0.1"

isort = "^5.13.2"
black = "^24.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
